도서상태 등록 기능 구현 (#136)
* featrue: BookStateEnum 한글이름 추가

* feature: 도서상태 요청 dto 추가

* fix: MehtodArgumentNotValidException 메서드 수정

* chore: 도서상태 id 타입 변경 및 파일 위치 변경

* feature: 도서상태 repository 추가

* featrue: 도서상태 register 서비스 추가

* feature: 도서상태 register 컨트롤러 추가

* refactor: 도서상태 정적 팩토리 메서드로 변경

* feat: 도서상태 등록 repository 테스트 추가

* feat: 도서상태 등록 service 테스트 추가

* feat: 도서상태 등록 controller 테스트 추가

* fix: 도서상태 등록 기능 X-USER-ID 누락 수정

* fix: 도서상태 등록 컨트롤러 스웨거 추가

* fix: 도서상태 등록 커스텀 예외처리 추가

+++ src/main/java/shop/nuribooks/books/dto/book/BookRegisterRes.java
@@ -9,7 +9,7 @@ import lombok.Getter;
 @Builder
 public class BookRegisterRes {
 	private Long id;
-	private Long stateId;
+	private Integer stateId;
 	private Long publisherId;
 	private String title;
 	private String thumbnailImageUrl;
